New Matrix Report
This report is a Matrix report that will loop through every plan you have loaded in the matrix view. It will print a report for every person who is scheduled to the plan based on the team(s) you configure via the variables. You can also have a custom sort order since the default is alphabetical. The last part of the report will be a roster of everyone scheduled for all of the plans in the matrix.